# BackEnd/docs/swagger.yaml
basePath: /
definitions:
  controllers.ToggleUserStatusRequest:
    properties:
      is_active:
        type: boolean
    type: object
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  models.SignupRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  models.UpdateMongoURIRequest:
    properties:
      database_name:
        type: string
      mongodb_uri:
        type: string
    required:
    - database_name
    - mongodb_uri
    type: object
host: localhost:8080
info:
  contact:
    email: support@schemacraft.com
    name: SchemaCraft Support
    url: http://www.schemacraft.com/support
  description: Dynamic Schema API Builder - Create APIs on the fly with custom schemas
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SchemaCraft API
  version: "1.0"
paths:
  /admin/stats:
    get:
      description: Get overall platform statistics
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get platform statistics (Admin only)
      tags:
      - admin
  /admin/users:
    get:
      description: Get all registered users with pagination
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get all users (Admin only)
      tags:
      - admin
  /admin/users/{id}:
    get:
      description: Get detailed user information by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get user by ID (Admin only)
      tags:
      - admin
  /admin/users/{id}/revoke-api-key:
    post:
      description: Revoke a user's API key
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Revoke user API key (Admin only)
      tags:
      - admin
  /admin/users/{id}/toggle-status:
    put:
      consumes:
      - application/json
      description: Activate or deactivate a user account
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Active status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.ToggleUserStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Toggle user active status (Admin only)
      tags:
      - admin
...
  /auth/me:
    get:
      description: Get current authenticated user information
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/mongodb-uri:
    put:
      consumes:
      - application/json
      description: Update user's custom MongoDB connection URI
      parameters:
      - description: MongoDB URI data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateMongoURIRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Update MongoDB URI
      tags:
      - auth
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: User signin
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Signup data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
        "400":
          description: Bad Request
        "409":
          description: Conflict
        "500":
          description: Internal Server Error
      summary: User signup
      tags:
      - auth
...
securityDefinitions:
  ApiKeyAuth:
    description: API Key for dynamic API access
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Bearer token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"